+++ cli/src/pcluster/config/validators.py
@@ -90,7 +90,7 @@ FSX_MESSAGES = {
 }
 
 FSX_SUPPORTED_ARCHITECTURES_OSES = {
-    "x86_64": ["centos7", "centos8", "ubuntu1604", "ubuntu1804", "alinux", "alinux2"],
+    "x86_64": ["centos7", "centos8", "ubuntu1804", "alinux2"],
     "arm64": ["ubuntu1804", "alinux2", "centos8"],
 }
 
@@ -1135,19 +1135,6 @@ def architecture_os_validator(param_key, param_value, pcluster_config):
         )
 
     return errors, warnings
-
-
-def base_os_validator(param_key, param_value, pcluster_config):
-    warnings = []
-
-    eol_2020 = ["alinux"]
-    if param_value in eol_2020:
-        warnings.append(
-            "The operating system you are using ({0}) will reach end-of-life in late 2020. It will be deprecated in "
-            "future releases of ParallelCluster".format(param_value)
-        )
-
-    return [], warnings
 
 
 def tags_validator(param_key, param_value, pcluster_config):