+++ tests/integration-tests/tests/common/utils.py
@@ -21,18 +21,9 @@ from utils import get_instance_info
 LOGGER = logging.getLogger(__name__)
 
 OS_TO_OFFICIAL_AMI_NAME_OWNER_MAP = {
-    "alinux": {"name": "amzn-ami-hvm-*.*.*.*-*-gp2", "owners": ["amazon"]},
     "alinux2": {"name": "amzn2-ami-hvm-*.*.*.*-*-gp2", "owners": ["amazon"]},
     "centos7": {"name": "CentOS 7.*", "owners": ["125523088429"]},
     "centos8": {"name": "CentOS 8.*", "owners": ["125523088429", "247102896272"]},
-    "ubuntu1404": {
-        "name": "ubuntu/images/hvm-ssd/ubuntu-trusty-14.04-*-server-*",
-        "owners": ["099720109477", "513442679011", "837727238323"],
-    },
-    "ubuntu1604": {
-        "name": "ubuntu/images/hvm-ssd/ubuntu-xenial-16.04-*-server-*",
-        "owners": ["099720109477", "513442679011", "837727238323"],
-    },
     "ubuntu1804": {
         "name": "ubuntu/images/hvm-ssd/ubuntu-bionic-18.04-*-server-*",
         "owners": ["099720109477", "513442679011", "837727238323"],
@@ -41,10 +32,8 @@ OS_TO_OFFICIAL_AMI_NAME_OWNER_MAP = {
 
 # Remarkable AMIs are latest deep learning base AMI and FPGA developer AMI without pcluster infrastructure
 OS_TO_REMARKABLE_AMI_NAME_OWNER_MAP = {
-    "alinux": {"name": "Deep Learning Base AMI (Amazon Linux)*", "owners": ["amazon"]},
     "alinux2": {"name": "Deep Learning Base AMI (Amazon Linux 2)*", "owners": ["amazon"]},
     "centos7": {"name": "FPGA Developer AMI*", "owners": ["679593333241"]},
-    "ubuntu1604": {"name": "Deep Learning Base AMI (Ubuntu 16.04)*", "owners": ["amazon"]},
     "ubuntu1804": {"name": "Deep Learning Base AMI (Ubuntu 18.04)*", "owners": ["amazon"]},
 }
 
@@ -52,11 +41,9 @@ OS_TO_REMARKABLE_AMI_NAME_OWNER_MAP = {
 PCLUSTER_AMI_OWNERS = ["amazon", "self"]
 # Pcluster AMIs are latest ParallelCluster official AMIs that align with cli version
 OS_TO_PCLUSTER_AMI_NAME_OWNER_MAP = {
-    "alinux": {"name": "amzn-hvm-x86_64-*", "owners": PCLUSTER_AMI_OWNERS},
     "alinux2": {"name": "amzn2-hvm-*-*", "owners": PCLUSTER_AMI_OWNERS},
     "centos7": {"name": "centos7-hvm-x86_64-*", "owners": PCLUSTER_AMI_OWNERS},
     "centos8": {"name": "centos8-hvm-x86_64-*", "owners": PCLUSTER_AMI_OWNERS},
-    "ubuntu1604": {"name": "ubuntu-1604-lts-hvm-x86_64-*", "owners": PCLUSTER_AMI_OWNERS},
     "ubuntu1804": {"name": "ubuntu-1804-lts-hvm-*-*", "owners": PCLUSTER_AMI_OWNERS},
 }