+++ util/generate-ami-list.py
@@ -29,11 +29,9 @@ from common import PARTITION_TO_MAIN_REGION, PARTITIONS
 
 DISTROS = OrderedDict(
     [
-        ("alinux", "amzn"),
         ("alinux2", "amzn2"),
         ("centos7", "centos7"),
         ("centos8", "centos8"),
-        ("ubuntu1604", "ubuntu-1604"),
         ("ubuntu1804", "ubuntu-1804"),
     ]
 )