+++ CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG
 
 **CHANGES**
 
+- Ubuntu 16.04 is no longer supported.
+- Amazon Linux is no longer supported.
 - Upgrade Slurm to version 20.11.4.
   - Add new SlurmctldParameters, power_save_min_interval=30, so power actions will be processed every 30 seconds
   - Specify instance GPU model as GRES GPU Type in gres.conf, instead of previous hardcoded value for all GPU, Type=tesla